Split DTOs within the dto folder.
