+++ backend/internal/controller/dto/house_request.go
@@ -1,0 +1,8 @@
+package dto
+
+type HouseRequest string
+
+func NewHouseRequest(name string) *HouseRequest {
+	request := HouseRequest(name)
+	return &request
+}
