+++ backend/internal/controller/dto/house_response.go
@@ -1,0 +1,13 @@
+package dto
+
+type HousesResponse struct {
+	ID   int    `json:"id"`
+	Name string `json:"name"`
+}
+
+func NewHousesResponse(id int, name string) *HousesResponse {
+	return &HousesResponse{
+		ID:   id,
+		Name: name,
+	}
+}
