+++ backend/internal/controller/dto/workflow_ui_request.go
@@ -1,0 +1,13 @@
+package dto
+
+type WorkflowUIRequest struct {
+	Nodes []NodeRequest `json:"nodes"`
+	Edges []EdgeRequest `json:"edges"`
+}
+
+func NewWorkflowUIRequest(nodes []NodeRequest, edges []EdgeRequest) *WorkflowUIRequest {
+	return &WorkflowUIRequest{
+		Nodes: nodes,
+		Edges: edges,
+	}
+}
