+++ backend/internal/controller/dto/workflow_ui_response.go
@@ -1,0 +1,13 @@
+package dto
+
+type WorkflowUIResponse struct {
+	Nodes []NodeResponse `json:"nodes"`
+	Edges []EdgeResponse `json:"edges"`
+}
+
+func NewWorkflowUIResponse(nodes []NodeResponse, edges []EdgeResponse) *WorkflowUIResponse {
+	return &WorkflowUIResponse{
+		Nodes: nodes,
+		Edges: edges,
+	}
+}
