+++ backend/internal/controller/dto/workflow_with_ui_request.go
@@ -1,0 +1,15 @@
+package dto
+
+type WorkflowWithUIRequest struct {
+	Workflow   Workflow          `json:"workflow"`
+	WorkflowUI WorkflowUIRequest `json:"workflow_ui"`
+	Relays     Relays            `json:"relays"`
+}
+
+func NewWorkflowWithUIRequest(workflow Workflow, workflowUI WorkflowUIRequest, relays Relays) *WorkflowWithUIRequest {
+	return &WorkflowWithUIRequest{
+		Workflow:   workflow,
+		WorkflowUI: workflowUI,
+		Relays:     relays,
+	}
+}
